% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_approximate_stretch}
\alias{get_approximate_stretch}
\title{Get approximate stretch factor}
\usage{
get_approximate_stretch(data, reference = "ref", query = "query")
}
\arguments{
\item{data}{Input data frame, either containing all replicates of gene expression or not.}

\item{reference}{Accession name of reference data.}

\item{query}{Accession name of query data.}
}
\value{
This function returns an estimation of a stretch factor for registering the data.
}
\description{
\code{get_approximate_stretch()} is a function to get a stretch factor
estimation given input data. This function will take the time point ranges
of both reference and query data and compare them to estimate the stretch
factor.
}
